backend/requirements.txt
# Core
Django>=5.0,<6
djangorestframework>=3.15,<4
django-cors-headers>=4.3,<5

# Config & server
python-dotenv>=1.0,<2
gunicorn>=22.0,<24
whitenoise>=6.7,<7

# Data & async (used in later phases; pinned for reproducible builds)
pymongo>=4.6,<5
celery[redis]>=5.4,<6
redis>=5.0,<6

# Webhooks / HTTP clients (phase 2+)
httpx>=0.27,<1
